fix: statusbar update on theme change
