+++ mobile/react-native-interview/src/components/__tests__/filterHeaderButton.test.tsx
@@ -1,17 +1,20 @@
 import React from 'react';
 import { render } from '@testing-library/react-native';
 import FilterHeaderButton from '../buttons/filterHeaderButton';
-import { ThemeProvider } from '@react-navigation/native';
+import { ThemeProvider as NavigationThemeProvider } from '@react-navigation/native';
 import { ArtCrimeFilters } from '../../types/artCrime';
 import { mockTheme } from '../../test-utils/mocks/theme';
+import { ThemeProvider } from '@/context/themeContext';
 
 describe('FilterHeaderButton', () => {
   it('renders correctly with no filters', () => {
     const mockFilters: ArtCrimeFilters = {};
 
     const { toJSON } = render(
-      <ThemeProvider value={mockTheme}>
-        <FilterHeaderButton filters={mockFilters} onPress={() => {}} />
+      <ThemeProvider>
+        <NavigationThemeProvider value={mockTheme}>
+          <FilterHeaderButton filters={mockFilters} onPress={() => {}} />
+        </NavigationThemeProvider>
       </ThemeProvider>,
     );
 
@@ -26,8 +29,10 @@ describe('FilterHeaderButton', () => {
     };
 
     const { toJSON } = render(
-      <ThemeProvider value={mockTheme}>
-        <FilterHeaderButton filters={mockFilters} onPress={() => {}} />
+      <ThemeProvider>
+        <NavigationThemeProvider value={mockTheme}>
+          <FilterHeaderButton filters={mockFilters} onPress={() => {}} />
+        </NavigationThemeProvider>
       </ThemeProvider>,
     );
