--- mobile/react-native-interview/src/navigation/navigationWrapper.tsx
@@ -1,13 +1,0 @@
-import { NavigationContainer } from '@react-navigation/native';
-import { useTheme } from '../context/themeContext';
-import RootNavigator from './rootNavigator';
-
-export default function NavigationWrapper() {
-  const { theme } = useTheme();
-
-  return (
-    <NavigationContainer theme={theme}>
-      <RootNavigator />
-    </NavigationContainer>
-  );
-}
