+++ mobile/react-native-interview/src/screens/_tabs/home/components/homeScreenContent.tsx
@@ -1,7 +1,7 @@
 import { View, StyleSheet } from 'react-native';
 import { ArtCrimeQueryParams } from '@/types/artCrime';
 import LoadingOverlay from '@/components/loadingOverlay';
-import InfinityGalleryList from '@/components/gallery/infinityGalleryList';
+import GalleryList from '@/components/gallery/galleryList';
 import { getErrorMessage } from '@/utils/getErrorMessage';
 import InlineError from '@/components/inlineError';
 import { useArtCrimesList } from '@/hooks/useArtCrimesList';
@@ -23,7 +23,7 @@ export default function HomeScreenContent({ queryParams }: HomeScreenContentProp
     );
   return (
     <View style={styles.container}>
-      <InfinityGalleryList
+      <GalleryList
         items={items || []}
         isFetchingNextPage={isFetchingNextPage}
         onEndReached={loadMore}
